Consolidated action selection decisions into a single class in the server
