+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/decisions/ActionSelectionDecision.java
@@ -1,55 +1,77 @@
 package com.gempukku.stccg.decisions;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+import com.gempukku.stccg.actions.Action;
 import com.gempukku.stccg.actions.TopLevelSelectableAction;
-import com.gempukku.stccg.cards.CardNotFoundException;
-import com.gempukku.stccg.cards.physicalcard.PhysicalCard;
+import com.gempukku.stccg.common.AwaitingDecisionType;
 import com.gempukku.stccg.common.DecisionResultInvalidException;
 import com.gempukku.stccg.game.DefaultGame;
+import com.gempukku.stccg.game.InvalidGameLogicException;
 import com.gempukku.stccg.player.Player;
 
+import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
-public abstract class ActionSelectionDecision extends ActionDecision {
+public abstract class ActionSelectionDecision extends AbstractAwaitingDecision {
+
+    final List<TopLevelSelectableAction> _actions;
+
+    @JsonProperty("context")
+    private final DecisionContext _context;
+
+    @JsonProperty("cardIds")
+    protected final String[] _cardIds;
+
+    private final String[] _actionTexts;
 
     @JsonProperty("min")
-    private final int _min = 1;
+    private final int _minActions;
     @JsonProperty("max")
-    private final int _max = 1;
+    private static final int MAX_SELECTIONS = 1;
 
     public ActionSelectionDecision(Player player, DecisionContext context, List<TopLevelSelectableAction> actions,
-                                   DefaultGame cardGame)
-            throws CardNotFoundException {
-        super(player, context, actions, cardGame);
-        setBlueprintIds(cardGame);
-        setParam("blueprintId", _blueprintIds.toArray(new String[0]));
-        setParam("imageUrl", getImageUrls(cardGame));
-        setParam("cardId", _cardIds);
+                                   DefaultGame cardGame, boolean required) {
+        super(player, context, AwaitingDecisionType.ACTION_CHOICE, cardGame);
+        _context = context;
+        _actions = actions;
+        _actionTexts = getActionTexts(cardGame);
+        _cardIds = getCardIds();
+        _minActions = required ? 1 : 0;
     }
 
-    protected void setBlueprintIds(DefaultGame cardGame) throws CardNotFoundException {
-        _blueprintIds.clear();
-        for (TopLevelSelectableAction action : _actions) {
-            int cardId = action.getCardIdForActionSelection();
-            PhysicalCard physicalCard = cardGame.getCardFromCardId(cardId);
-            _blueprintIds.add(String.valueOf(physicalCard.getBlueprintId()));
-        }
+    public String getElementType() { return "ACTION"; }
+
+    public String[] getActionTexts(DefaultGame game) {
+        String[] result = new String[_actions.size()];
+        for (int i = 0; i < result.length; i++)
+            try {
+                result[i] = _actions.get(i).getActionSelectionText(game);
+            } catch(InvalidGameLogicException exp) {
+                result[i] = "Select action";
+            }
+        return result;
     }
 
-    private String[] getImageUrls(DefaultGame cardGame) throws CardNotFoundException {
-        String[] images = new String[_actions.size()];
-        for (int i = 0; i < images.length; i++) {
-            int cardId = _actions.get(i).getCardIdForActionSelection();
-            PhysicalCard physicalCard = cardGame.getCardFromCardId(cardId);
-            images[i] = (physicalCard == null) ? "rules" : physicalCard.getImageUrl();
-        }
-        return images;
+    public String[] getCardIds() {
+        String[] result = new String[_actions.size()];
+        for (int i = 0; i < result.length; i++)
+            result[i] = String.valueOf(_actions.get(i).getCardIdForActionSelection());
+        return result;
+    }
+
+    public List<TopLevelSelectableAction> getActions() { return _actions; }
+
+    public void decisionMade(Action action) throws DecisionResultInvalidException {
+        if (action instanceof TopLevelSelectableAction && _actions.contains(action))
+            decisionMade(String.valueOf(_actions.indexOf(action)));
+        else throw new DecisionResultInvalidException("Action not found in ActionDecision");
     }
 
     protected TopLevelSelectableAction getSelectedAction(String result) throws DecisionResultInvalidException {
         if (result.isEmpty())
-            throw new DecisionResultInvalidException();
-
+            return null;
         try {
             int actionIndex = Integer.parseInt(result);
             if (actionIndex < 0 || actionIndex >= _actions.size())
@@ -61,10 +83,17 @@ public abstract class ActionSelectionDecision extends ActionDecision {
         }
     }
 
-    public String[] getCardIds() {
-        String[] result = new String[_actions.size()];
-        for (int i = 0; i < result.length; i++) {
-            result[i] = String.valueOf(_actions.get(i).getCardIdForActionSelection());
+    @JsonProperty("displayedCards")
+    protected List<Map<Object, Object>> getDisplayedCards() {
+        List<Map<Object, Object>> result = new ArrayList<>();
+        for (int i = 0; i < _cardIds.length; i++) {
+            Map<Object, Object> mapToAdd = new HashMap<>();
+            mapToAdd.put("cardId", _cardIds[i]);
+            mapToAdd.put("actionId", String.valueOf(i));
+            mapToAdd.put("actionText", _actionTexts[i]);
+            mapToAdd.put("actionType", _actions.get(i).getActionType());
+            mapToAdd.put("selectable", "true");
+            result.add(mapToAdd);
         }
         return result;
     }
