+++ gemp-module/gemp-stccg-logic/src/main/java/com/gempukku/stccg/processes/st1e/ST1EFacilitySeedPhaseProcess.java
@@ -8,12 +8,10 @@ import com.gempukku.stccg.actions.draw.DrawMultipleCardsUnrespondableAction;
 import com.gempukku.stccg.cards.physicalcard.PhysicalCard;
 import com.gempukku.stccg.common.DecisionResultInvalidException;
 import com.gempukku.stccg.common.filterable.Zone;
-import com.gempukku.stccg.decisions.CardActionSelectionDecision;
+import com.gempukku.stccg.decisions.ActionSelectionDecision;
 import com.gempukku.stccg.decisions.DecisionContext;
 import com.gempukku.stccg.game.DefaultGame;
 import com.gempukku.stccg.game.InvalidGameLogicException;
-import com.gempukku.stccg.game.ST1EGame;
-import com.gempukku.stccg.gamestate.ST1EGameState;
 import com.gempukku.stccg.player.Player;
 import com.gempukku.stccg.player.PlayerNotFoundException;
 import com.gempukku.stccg.player.PlayerOrder;
@@ -22,7 +20,6 @@ import com.gempukku.stccg.processes.StartOfTurnGameProcess;
 
 import java.beans.ConstructorProperties;
 import java.util.Collection;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -45,8 +42,8 @@ public class ST1EFacilitySeedPhaseProcess extends ST1EGameProcess {
             _consecutivePasses++;
         } else {
             cardGame.getUserFeedback().sendAwaitingDecision(
-                    new CardActionSelectionDecision(currentPlayer, DecisionContext.SELECT_PHASE_ACTION,
-                            playableActions, cardGame) {
+                    new ActionSelectionDecision(currentPlayer, DecisionContext.SELECT_PHASE_ACTION,
+                            playableActions, cardGame, false) {
                         @Override
                         public void decisionMade(String result) throws DecisionResultInvalidException {
                             try {
